agents/langchain-agent-be/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# LangChain 1.0 and MCP
langchain>=1.0.0
langchain-openai>=0.3.32
langchain-core>=0.3.75
langchain-mcp-adapters>=0.1.0

# HTTP client (for MCP SSE transport)
httpx>=0.25.0
